Validate that posted attachments are not None
When submitting a new document without file attachments, the serializer
tried to check for file limits and pop the data from the original
validated data. Because no files were uploaded, the value was None, and
it was failing the check for max uploads and later when popping them
from the data.

Closes Sentry error ATV-API-8