deps: remove drf-oidc-auth
drf-oidc-auth is not used so it can be removed.
This package removal caused some import error since pycryptodomex was
removed from the requirements but the pycryptodome (without x) is still
used by django-searchable-encrypted-fields but it has different import
name so made the changes for import clauses.
Note that pycryptodomex should be PyCrypto compatible/not clashing the
namespaces unlike the pycryptodome. This project does not require
pycrypto and should be safe.

Refs: ATV-243